@@ -146,12 +146,10 @@ public partial class Actor : GameObject {
 	float createRpcTime;
 
 	public bool splashable;
-	private Anim _waterWade = null!;
+	private Anim? _waterWade = null;
 	public Anim waterWade {
 		get {
-			if (_waterWade == null) {
-				_waterWade = new Anim(pos, "wade", 1, null, false);
-			}
+			_waterWade ??= new Anim(pos, "wade", 1, null, false);
 			return _waterWade;
 		}
 	}
@@ -1458,6 +1456,9 @@ public partial class Actor : GameObject {
 
 
 	public SoundWrapper? playSound(string soundKey, bool forcePlay = false, bool sendRpc = false) {
+		if (soundKey == "") {
+			return null;
+		}
 		soundKey = soundKey.ToLowerInvariant();
 		if (!Global.soundBuffers.ContainsKey(soundKey)) {
 			throw new Exception($"Attempted playing missing sound with name \"{soundKey}\"");
