@@ -239,7 +239,7 @@ public partial class Character : Actor, IDamagable {
 		bool isVisible, ushort? netId, bool ownedByLocalPlayer,
 		bool isWarpIn = true, int? heartTanks = null, bool isATrans = false
 	) : base(
-		null!, new Point(x, y), netId, ownedByLocalPlayer, addToLevel: true
+		"", new Point(x, y), netId, ownedByLocalPlayer, addToLevel: true
 	) {
 		//hasStateMachine = true;
 		slideOnIce = true;
@@ -720,8 +720,7 @@ public partial class Character : Actor, IDamagable {
 
 	// For terrain collision.
 	public override Collider? getTerrainCollider() {
-		Collider? overrideGlobalCollider = null;
-		if (spriteToColliderMatch(sprite.name, out overrideGlobalCollider)) {
+		if (spriteToColliderMatch(sprite.name, out Collider? overrideGlobalCollider)) {
 			return overrideGlobalCollider;
 		}
 		if (physicsCollider == null) {
@@ -802,9 +801,7 @@ public partial class Character : Actor, IDamagable {
 		if (grounded && !isDashing) {
 			dashedInAir = 0;
 		}
-		if (ai != null) {
-			ai.preUpdate();
-		}
+		ai?.preUpdate();
 	}
 
 	public override void onCollision(CollideData other) {
@@ -1181,9 +1178,7 @@ public partial class Character : Actor, IDamagable {
 			usedSubtank = null;
 		}
 
-		if (ai != null) {
-			ai.update();
-		}
+		ai?.update();
 
 		if (slideVel != 0) {
 			slideVel = Helpers.toZero(slideVel, speedMul * 0.1f, Math.Sign(slideVel));
@@ -1391,7 +1386,7 @@ public partial class Character : Actor, IDamagable {
 			if (player.input.isPressed(Control.Jump, player) && canJump()) {
 				vel.y = -getJumpPower();
 				isDashing = (
-					isDashing || player.dashPressed(out string dashControl) && canDash()
+					isDashing || player.dashPressed(out _) && canDash()
 				);
 				if (isDashing) {
 					dashedInAir++;
@@ -1624,8 +1619,7 @@ public partial class Character : Actor, IDamagable {
 			int chargeType = 0;
 			if (!sprite.name.Contains("ra_hide")) {
 				int level = getDisplayChargeLevel();
-				var renderGfx = RenderEffectType.ChargeBlue;
-				renderGfx = level switch {
+				RenderEffectType renderGfx = level switch {
 					1 => RenderEffectType.ChargeBlue,
 					2 => RenderEffectType.ChargeYellow,
 					3 when (chargeType == 2) => RenderEffectType.ChargeOrange,
@@ -2116,11 +2110,9 @@ public partial class Character : Actor, IDamagable {
 			changeSprite(getSprite(newState.shootSpriteEx), true);
 		} else {
 			string spriteName = sprite.name;
-			string targetSprite = newState.sprite;
 			if (newState.sprite == newState.transitionSprite &&
 				!Global.sprites.ContainsKey(getSprite(newState.transitionSprite))
 			) {
-				targetSprite = newState.defaultSprite;
 				newState.sprite = newState.defaultSprite;
 			}
 			if (Global.sprites.ContainsKey(getSprite(newState.sprite))) {
@@ -2198,11 +2190,11 @@ public partial class Character : Actor, IDamagable {
 				0, pos.x + x, pos.y + y + yOff, xDir, 1, null, 1, 1, 1, zIndex + 1
 			);
 		}
-		List<Player> nonSpecPlayers = Global.level.nonSpecPlayers();
+		//List<Player> nonSpecPlayers = Global.level.nonSpecPlayers();
 
-		bool drawCursorChar = player.isMainPlayer && (
+		/*bool drawCursorChar = player.isMainPlayer && (
 			Global.level.is1v1() || Global.level.server.fixedCamera
-		);
+		);*/
 
 		bool shouldDrawName = false;
 		bool shouldDrawHealthBar = false;
@@ -2365,7 +2357,6 @@ public partial class Character : Actor, IDamagable {
 				Alignment.Center, true, depth: ZIndex.HUD
 			);
 			if (ai != null) {
-				var charTarget = ai.target as Character;
 				Fonts.drawText(
 					FontType.Grey, "dest:" + ai.aiState.getDestNodeName(),
 					textPosX, textPosY -= 10, Alignment.Center, true, depth: ZIndex.HUD
@@ -2382,7 +2373,7 @@ public partial class Character : Actor, IDamagable {
 					FontType.Grey, ai.aiState.GetType().ToString().RemovePrefix("MMXOnline."),
 					textPosX, textPosY -= 10, Alignment.Center, true, depth: ZIndex.HUD
 				);
-				if (charTarget != null) {
+				if (ai.target is Character charTarget) {
 					Fonts.drawText(
 						FontType.Grey, "target:" + charTarget?.name, textPosX, textPosY -= 10,
 						Alignment.Center, true, depth: ZIndex.HUD
@@ -2489,7 +2480,7 @@ public partial class Character : Actor, IDamagable {
 
 		if (charState is GenericStun gst) {
 			bool hasDrawn = false;
-			List<int> iconsToDraw = new();
+			List<int> iconsToDraw = [];
 			if (crystalizedTime > 0) {
 				drawStatusBar(crystalizedTime, gst.getTimerFalloff(), crystalizedMaxTime, new Color(247, 206, 247));
 				deductLabelY(5);
@@ -2804,8 +2795,7 @@ public partial class Character : Actor, IDamagable {
 			crystalizedTime = 0; // Dash to destroy crystal
 		}
 
-		var inRideArmor = charState as InRideArmor;
-		if (inRideArmor != null && inRideArmor.crystalizeTime > 0) {
+		if (charState is InRideArmor inRideArmor && inRideArmor.crystalizeTime > 0) {
 			if (weaponIndex == 20 && damage > 0) inRideArmor.crystalizeTime = 0;   //Dash to destroy crystal
 			inRideArmor.checkCrystalizeTime();
 		}
@@ -3076,9 +3066,7 @@ public partial class Character : Actor, IDamagable {
 			} else if (Global.level.gameMode.level.is1v1()) {
 				// In 1v1 the other player should always be considered a killer to prevent suicide
 				var otherPlayer = Global.level.nonSpecPlayers().Find(p => p.id != player.id);
-				if (otherPlayer != null) {
-					otherPlayer.addKill();
-				}
+				otherPlayer?.addKill();
 			}
 
 			if (assister != null && assister != player && assister != Player.stagePlayer) {
@@ -3400,7 +3388,7 @@ public partial class Character : Actor, IDamagable {
 			shape = shape.clone(0, MathF.Abs(shape.maxY) + 1);
 		}
 
-		var collision = Global.level.checkCollisionShape(shape, new List<GameObject>() { rideArmor });
+		var collision = Global.level.checkCollisionShape(shape, [rideArmor]);
 		if (collision?.gameObject is not Wall) {
 			return true;
 		}
