@@ -358,7 +358,8 @@ public class StormEGustProj : Projectile {
 
 #region states
 public class SEagleMState : MaverickState {
-	public StormEagle StormEagleed = null!;
+	public StormEagle stormEagleed = null!;
+
 	public SEagleMState(
 		string sprite, string transitionSprite = ""
 	) : base(
@@ -368,7 +369,7 @@ public class SEagleMState : MaverickState {
 
 	public override void onEnter(MaverickState oldState) {
 		base.onEnter(oldState);
-		StormEagleed = maverick as StormEagle ?? throw new NullReferenceException();
+		stormEagleed = maverick as StormEagle ?? throw new NullReferenceException();
 	}
 }
 public class StormEDiveState : SEagleMState {
@@ -457,7 +458,7 @@ public class StormEGustState : SEagleMState {
 
 	public override void update() {
 		base.update();
-		if (StormEagleed == null) return;
+		if (stormEagleed == null) return;
 
 		soundTime += Global.spf;
 		if (soundTime > 0.4f) {
@@ -471,7 +472,7 @@ public class StormEGustState : SEagleMState {
 			float randX = maverick.pos.x + maverick.xDir * Helpers.randomRange(0, 65);
 			Point pos = new Point(randX, maverick.pos.y);
 			new StormEGustProj(
-				pos, maverick.xDir, StormEagleed, player,
+				pos, maverick.xDir, stormEagleed, player,
 				player.getNextActorNetId(), rpc: true
 			);
 		}
@@ -500,13 +501,13 @@ public class StormEEggState : SEagleMState {
 
 	public override void update() {
 		base.update();
-		if (StormEagleed == null) return;
+		if (stormEagleed == null) return;
 
 		if (maverick.frameIndex == 3 && !once) {
 			once = true;
 			new StormEEggProj(
-				StormEagleed.getFirstPOI() ?? StormEagleed.getCenterPos(), maverick.xDir,
-				StormEagleed, player, player.getNextActorNetId(), rpc: true
+				stormEagleed.getFirstPOI() ?? stormEagleed.getCenterPos(), maverick.xDir,
+				stormEagleed, player, player.getNextActorNetId(), rpc: true
 			);
 		}
 
@@ -524,14 +525,14 @@ public class StormEAirShootState : SEagleMState {
 
 	public override void update() {
 		base.update();
-		if (StormEagleed == null) return;
+		if (stormEagleed == null) return;
 
 		Point? shootPos = maverick.getFirstPOI();
 		if (!shotOnce && shootPos != null) {
 			shotOnce = true;
 			maverick.playSound("tornado", sendRpc: true);
 			new TornadoProj(
-				shootPos.Value, maverick.xDir, true, StormEagleed,
+				shootPos.Value, maverick.xDir, true, stormEagleed,
 				player, player.getNextActorNetId(), rpc: true
 			);
 		}
