@@ -61,7 +61,7 @@ public class TriadThunder : Weapon {
 public class TriadThunderProj : Projectile {
 	int state;
 	Character? character;
-	MegamanX mmx = null!;
+	MegamanX? mmx = null;
 	public List<TriadThunderBall> balls;
 	public TriadThunderProj(
 		Point pos, int xDir, int yDir, Actor owner, Player player, ushort netProjId, bool rpc = false
@@ -74,7 +74,9 @@ public class TriadThunderProj : Projectile {
 		damager.flinch = Global.miniFlinch;
 		projId = (int)ProjIds.TriadThunder;
 		character = player.character;
-		mmx = player.character as MegamanX ?? throw new NullReferenceException();
+		if (ownedByLocalPlayer) {
+			mmx = player.character as MegamanX;
+		}
 		destroyOnHit = false;
 		shouldShieldBlock = false;
 		shouldVortexSuck = false;
